Move popup and change charts animation
Charts animation is disabled so labels and error bars can appear. Bug with Recharts library
