+++ src/App.tsx
@@ -1,27 +1,24 @@
-import React, { useContext, useEffect, useState } from "react";
+import React, { useContext, useEffect } from "react";
+import { useMediaQuery } from "react-responsive";
 import { Redirect, Route, Switch } from "react-router-dom";
 import './App.css';
 import About from './components/pages/About';
+import CookiePolicy from "./components/pages/CookiePolicy";
 import Dashboard from "./components/pages/Dashboard/Dashboard";
 import Data from './components/pages/Data';
+import Insights from "./components/pages/insights/Insights";
 import PrivacyPolicy from './components/pages/PrivacyPolicy';
+import TermsOfUse from "./components/pages/TermsOfUse";
+import { CookieBanner } from "./components/shared/CookieBanner";
 import { NavBar } from "./components/shared/NavBar";
+import { mobileDeviceOrTabletWidth } from "./constants";
 import { AppContext } from "./context";
 import httpClient from "./httpClient";
 import { setLanguageType } from "./utils/translate/translateService";
-import CookiePolicy from "./components/pages/CookiePolicy";
-import TermsOfUse from "./components/pages/TermsOfUse";
-import Insights from "./components/pages/insights/Insights";
-import { CookieBanner } from "./components/shared/CookieBanner";
-import { Modal } from "semantic-ui-react";
-import Translate from "./utils/translate/translateService";
-import { useMediaQuery } from "react-responsive";
-import { mobileDeviceOrTabletWidth } from "./constants";
 
 function App() {
   const isMobileDeviceOrTablet = useMediaQuery({ maxDeviceWidth: mobileDeviceOrTabletWidth });
   const [{ language }, dispatch] = useContext(AppContext);
-  const [showModal, toggleModal] = useState(true);
   // DATA
   useEffect(() => {
     const api = new httpClient()
@@ -51,26 +48,11 @@ function App() {
     setLanguageType(language);
   }, [dispatch, language])
 
-  const closeModal = () => toggleModal(false);
-
-  const MobileInfoModal = () => (
-    <Modal className="modal" open={showModal} onClose={closeModal} closeIcon={{ style: { top: '1.0535rem', right: '1rem' }, color: 'black', name: 'close' }}>
-      <Modal.Content className="modal-content">
-        <div className={isMobileDeviceOrTablet ? "modal-text-mobile" : "modal-text"}>
-          <p>{Translate('InitInfoModalText', ['PartOne'])}</p>
-          <p>{Translate('InitInfoModalText', ['PartTwo'])}</p>
-          <p>{Translate('InitInfoModalText', ['PartThree'])}</p>
-        </div>
-      </Modal.Content>
-    </Modal>
-  )
-
   // ROUTING TABS
   return (
     <div className="App">
       <NavBar />
       <CookieBanner />
-      <MobileInfoModal/>
       <Switch>
         <Route path="/About">
           <About />
